pyproject.toml
[project]
name = "parser-auto-ria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bs4>=0.0.2",
    "fake-useragent>=2.2.0",
    "fastapi>=0.136.1",
    "greenlet>=3.4.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.14.0",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
